+++ src/extension.ts
@@ -80,7 +80,7 @@ async function activate(context: vscode.ExtensionContext) {
       testRunner.onTestTree(testController, tree);
     },
   });
-  clientProvider.init();
+  await clientProvider.init();
 
   lsp.registerGlobally(clientProvider);
 
@@ -293,6 +293,11 @@ async function activate(context: vscode.ExtensionContext) {
         );
     },
     togglePrettyPrint: eval.togglePrettyPrint,
+    activateCalva: () => {
+      return new Promise((resolve, _reject) => {
+        resolve(true);
+      });
+    },
   };
 
   function registerCalvaCommand([command, callback]) {