+++ CHANGELOG.md
@@ -4,9 +4,21 @@ Changes to Calva.
 
 ## [Unreleased]
 
-- [Sort aliases for deps.edn projects](https://github.com/BetterThanTomorrow/calva/issues/2035)
+## [2.0.329] - 2023-02-01
 
-## [2.0.327] - 2023-01-26
+- [Sort aliases for deps.edn projects](https://github.com/BetterThanTomorrow/calva/issues/2035)
+- [Sort pre-selected project at the top in REPL connect menu](https://github.com/BetterThanTomorrow/calva/issues/2043)
+- Fix: [Indenter and formatter not in agreement about some forms](https://github.com/BetterThanTomorrow/calva/issues/2032)
+- Fix: [Regressions introduced with clojure-lsp multi-project support in 2.0.327](https://github.com/BetterThanTomorrow/calva/issues/2041)
+- Fix: [Can't use add require feature after updating to 2.0.327 version](https://github.com/BetterThanTomorrow/calva/issues/2040)
+- Fix: [Formatting issues on backspace](https://github.com/BetterThanTomorrow/calva/issues/2038)
+- Calva development: [Test the built VSIX extension in CI](https://github.com/BetterThanTomorrow/calva/issues/2051)
+- Calva development, Fix: [We build Calva twice in CI](https://github.com/BetterThanTomorrow/calva/issues/2052)
+
+## [2.0.328] - 2023-01-27
+- Rollback of 2.0.327, first part of: [Regressions introduced with clojure-lsp multi-project support in 2.0.327](https://github.com/BetterThanTomorrow/calva/issues/2041)
+
+## [2.0.327] - 2023-01-27
 
 - [LSP support for multi-project and multi-workspace](https://github.com/BetterThanTomorrow/calva/pull/2020)
 - Fix: [Clojure-lsp caching in multi-project workspace](https://github.com/BetterThanTomorrow/calva/issues/934)